// types/product.type.ts
export interface IProduct {
    id: number;
    title: string;
    description: string;
    category: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    tags: Array<string>;
    brand: string;
    sku: string;
    weight: number;
    dimensions: {
      width: number;
      height: number;
      depth: number;
    };
    warrantyInformation: string;
    shippingInformation: string;
    availabilityStatus: string;
    reviews: [
      {
        rating: number;
        comment: string;
        date: string;
        reviewerName: string;
        reviewerEmail: string;
      },
      {
        rating: number;
        comment: string;
        date: string;
        reviewerName: string;
        reviewerEmail: string;
      },
      {
        rating: number;
        comment: string;
        date: string;
        reviewerName: string;
        reviewerEmail: string;
      }
    ];
    returnPolicy: string;
    minimumOrderQuantity: number;
    meta: {
      createdAt: string;
      updatedAt: string;
      barcode: number;
      qrCode: string;
    };
    images: string[];
    thumbnail: string;
  }
  